Add visualization and data utility functions
Added two new files: `utils.py` which includes helper functions for data manipulation and authentication, and `expenses.py` which includes visualization for the application. Also, plotly and plotly-express have been added to the requirements.txt for visualization purposes. A new code snippet for visualization of top consumed products was added in the `wrangling.ipynb`.